requirements.txt
django==3.2.4
djangorestframework==3.12.4
mysqlclient==2.0.3
drf-yasg==1.20.0
django-filter==2.4.0
